[~] panel rename
