+++ src/constants/index.js
@@ -1,7 +1,7 @@
-export { default as ADMIN_PANEL } from './adminPanel';
 export { default as aliases } from './aliases';
 export { default as GRAPH } from './graph';
 export { default as LOCAL_STORAGE } from './local-storage';
+export { default as ADMIN_PANEL } from './panel';
 export { default as PRIORITY } from './priority';
 export { default as RAW } from './raw/index';
 export { default as SOCIALS } from './socials';
