+++ src/constants/panel.js
