--- src/state/slices/adminPanel/index.js
@@ -1,3 +1,0 @@
-import adminPanelSlice from './adminPanel.slice';
-
-export default adminPanelSlice;
