+++ src/state/slices/index.js
@@ -1,6 +1,6 @@
-export { default as adminPanelSlice } from './adminPanel';
 export { default as experimentsSlice } from './experiments';
 export { default as jobsSlice } from './jobs';
+export { default as panelSlice } from './panel';
 export { default as scrollSlice } from './scroll';
 export { default as sidebarSlice } from './sidebar';
 export { default as socketSlice } from './socket';
