+++ src/state/slices/panel/index.js
@@ -1,0 +1,3 @@
+import panel from './panel.slice';
+
+export default panel;
