+++ src/state/slices/panel/panel.reducers.js
@@ -1,5 +1,5 @@
 export const setValue = (state, { payload }) => {
-  state.adminPanel.value = payload;
+  state.panel.value = payload;
 };
 
 export const toggle = state => {
